Refactor `requests` module
